move recv_packet
